+++ src/get-sass-vars.d.ts
@@ -1,0 +1,4 @@
+declare module "get-sass-vars"
+{
+    export {};
+}