+++ src/tests/Compilation/Presentation/ThemeCompiler.test.ts
@@ -39,6 +39,7 @@ export function ThemeCompilerTests(): void
                 {
                     Identifier: "bar",
                     DisplayName: {},
+                    Version: "",
                     InstallSet: {
                         Instructions: []
                     }