+++ src/tests/Compilation/Presentation/ThemeFileCompiler.test.ts
@@ -72,6 +72,7 @@ export function ThemeFileCompilerTests(): void
                 {
                     DisplayName: {},
                     Identifier: "foo.bar",
+                    Version: "",
                     InstallSet: {
                         Instructions: []
                     }